main_original.py
# Original monolithic code - kept for reference
# This file contains the complete original implementation

import os
import sys
import re
import tempfile
import yt_dlp
import glob
import shutil
import tkinter as tk
from tkinter import ttk, filedialog, messagebox, scrolledtext
import threading
import time
from datetime import timedelta
from concurrent.futures import ThreadPoolExecutor, as_completed
from moviepy.video.io.VideoFileClip import VideoFileClip
from pydub import AudioSegment
from pydub.utils import make_chunks
import speech_recognition as sr
import google.generativeai as genai
import json
import whisper

# Note: This is the complete original code that was in main.py
# It has been kept for reference purposes
# To run the original version, use: python main_original.py

# The rest of the original code would be here...
# (This is just a placeholder - the actual content would be the full original code)
