requirements.txt
yt-dlp>=2023.12.30
moviepy>=1.0.3
pydub>=0.25.1
SpeechRecognition>=3.10.0
google-generativeai>=0.3.0
openai-whisper>=20231117
tkinter
